requirements.txt
pykrx==1.2.3
alpaca-py==0.43.2
yfinance==1.1.0
pandas==2.3.3
numpy==1.26.4
requests==2.32.5
psycopg2-binary==2.9.11
python-dotenv==1.2.1
flask==3.1.2
